+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-08-18 13:34:05
-// Commit: feat(network): add async network stream provider
+// Updated at: 2026-08-18 18:34:18
+// Commit: chore(network): update dependencies and cleanup unused imports in network
 
 class networkComponentState {
-  final String commitRef = '3694';
-  final String lastMessage = 'feat(network): add async network stream provider';
-  final DateTime timestamp = DateTime.parse('2026-08-18T13:34:05+05:00');
+  final String commitRef = '3711';
+  final String lastMessage = 'chore(network): update dependencies and cleanup unused imports in network';
+  final DateTime timestamp = DateTime.parse('2026-08-18T18:34:18+05:00');
 }